requirements.txt
pandas==2.1.4
numpy==1.24.3
pyyaml==6.0.1
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0
streamlit==1.29.0
scipy==1.11.4
scikit-learn==1.3.2
sqlalchemy==2.0.23
psycopg2-binary==2.9.9